fix: cleanup config, move local mock provider to tests
